requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-WTF==1.1.1
mysql-connector-python==8.1.0
PyMySQL==1.1.0
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==2.3.7
email-validator==2.0.0.post2
cryptography==41.0.3
SQLAlchemy-Utils==0.41.1
breeze-connect==1.0.37
WTForms==3.0.1
